% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cforest.R
\name{cforest}
\alias{cforest}
\title{Conditional Random Forests
See \code{?partykit::cforest}.}
\usage{
cforest(
  formula,
  data,
  weights,
  subset,
  offset,
  cluster,
  strata,
  na.action = na.pass,
  control = partykit::ctree_control(teststat = "quad", testtype = "Univ", mincriterion
    = 0, saveinfo = FALSE, ...),
  ytrafo = NULL,
  scores = NULL,
  ntree = 500L,
  perturb = list(replace = FALSE, fraction = 0.632),
  mtry = ceiling(sqrt(nvar)),
  applyfun = NULL,
  cores = NULL,
  trace = FALSE,
  ...
)
}
\arguments{
\item{formula}{(see \code{?partykit::cforest})}

\item{data}{(see \code{?partykit::cforest})}

\item{weights}{(see \code{?partykit::cforest})}

\item{subset}{(see \code{?partykit::cforest})}

\item{offset}{(see \code{?partykit::cforest})}

\item{cluster}{(see \code{?partykit::cforest})}

\item{strata}{(see \code{?partykit::cforest})}

\item{na.action}{(see \code{?partykit::cforest})}

\item{control}{(see \code{?partykit::cforest})}

\item{ytrafo}{(see \code{?partykit::cforest})}

\item{scores}{(see \code{?partykit::cforest})}

\item{ntree}{(see \code{?partykit::cforest})}

\item{perturb}{(see \code{?partykit::cforest})}

\item{mtry}{(see \code{?partykit::cforest})}

\item{applyfun}{(see \code{?partykit::cforest})}

\item{cores}{(see \code{?partykit::cforest})}

\item{trace}{(see \code{?partykit::cforest})}

\item{...}{(see \code{?partykit::cforest})}
}
\description{
Conditional Random Forests
See \code{?partykit::cforest}.
}
