% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchObjectForValue.R
\name{deparseLayer}
\alias{deparseLayer}
\title{Get the first level structure of an R object}
\usage{
deparseLayer(obj)
}
\arguments{
\item{obj}{[`R object`] R object which is deparsed.}
}
\description{
Get the first level structure of an R object
}
\examples{
deparseLayer(iris$Sepal.Length)
}
