% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchObjectForValue.R
\name{deparseStructure}
\alias{deparseStructure}
\title{Deparse the whole structure of an R object}
\usage{
deparseStructure(obj, terminator = isFinalLayer)
}
\arguments{
\item{obj}{[`R object`] R object which is deparsed.}

\item{terminator}{[`function`] Function to check if a final layer is reached (default `terminator = isFinalLayer`).}
}
\description{
Deparse the whole structure of an R object
}
\examples{
obj = lm(Petal.Length ~ ., data = iris)
strcts = deparseStructure(obj)
}
