% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchObjectForValue.R
\name{findNumber}
\alias{findNumber}
\title{Find number in R object}
\usage{
findNumber(obj, number)
}
\arguments{
\item{obj}{\code{\link{R object}} R object which should be flatten.}

\item{number}{\code{\link{anything}} Value which which for which \code{obj} is searched through.}
}
\value{
\code{\link{list}} Containing messages whether the object was found or not.
}
\description{
Find number in R object
}
