% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchObjectForValue.R
\name{flattenObject}
\alias{flattenObject}
\title{Flatten R object}
\usage{
flattenObject(obj, patience = 10L)
}
\arguments{
\item{obj}{\code{\link{R object}} R object which should be flatten.}

\item{patience}{\code{\link{integer(1L)}} How many iterations needs the length of the old and new flatten object to be equal to stop the algorithm?}
}
\value{
\code{\link{list}} containing the flatten object.
}
\description{
Flatten R object
}
\examples{
obj = lm(Petal.Length ~ ., data = iris)
obj_flatten = flattenObject(obj)
}
