% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchObjectForValue.R
\name{isFinalLayer}
\alias{isFinalLayer}
\title{Check if a layer is the last one in the object structure}
\usage{
isFinalLayer(layer)
}
\arguments{
\item{layer}{[`character(1L)`]}
}
\description{
Check if a layer is the last one in the object structure
}
\examples{
isFinalLayer("iris$Sepal.Length[[140]]")
isFinalLayer("iris$Sepal.Length")
}
