% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cforest.R
\name{predict.cforest.nodat}
\alias{predict.cforest.nodat}
\title{Conditional Random Forests
See \code{?partykit::cforest}.}
\usage{
\method{predict}{cforest.nodat}(
  object,
  newdata = NULL,
  type = c("response", "prob", "weights", "node"),
  OOB = FALSE,
  FUN = NULL,
  simplify = TRUE,
  scale = TRUE,
  ...
)
}
\arguments{
\item{object}{(see \code{?partykit::predict.cforest})}

\item{newdata}{(see \code{?partykit::predict.cforest})}

\item{type}{(see \code{?partykit::predict.cforest})}

\item{OOB}{(see \code{?partykit::predict.cforest})}

\item{FUN}{(see \code{?partykit::predict.cforest})}

\item{simplify}{(see \code{?partykit::predict.cforest})}

\item{scale}{(see \code{?partykit::predict.cforest})}

\item{...}{(see \code{?partykit::predict.cforest})}
}
\description{
Conditional Random Forests
See \code{?partykit::cforest}.
}
