% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traforest.R
\name{predict.traforest.nodat}
\alias{predict.traforest.nodat}
\title{Transformation Forests
See \code{?trtf::predict.traforest}.}
\usage{
\method{predict}{traforest.nodat}(
  object,
  newdata,
  mnewdata = data.frame(1),
  K = 20,
  q = NULL,
  type = c("weights", "node", "coef", "trafo", "distribution", "survivor", "density",
    "logdensity", "hazard", "loghazard", "cumhazard", "quantile"),
  OOB = FALSE,
  simplify = FALSE,
  trace = FALSE,
  updatestart = FALSE,
  applyfun = NULL,
  cores = NULL,
  ...
)
}
\arguments{
\item{object}{(see \code{?trtf::predict.traforest})}

\item{newdata}{(see \code{?trtf::predict.traforest})}

\item{mnewdata}{(see \code{?trtf::predict.traforest})}

\item{K}{(see \code{?trtf::predict.traforest})}

\item{q}{(see \code{?trtf::predict.traforest})}

\item{type}{(see \code{?trtf::predict.traforest})}

\item{OOB}{(see \code{?trtf::predict.traforest})}

\item{simplify}{(see \code{?trtf::predict.traforest})}

\item{trace}{(see \code{?trtf::predict.traforest})}

\item{updatestart}{(see \code{?trtf::predict.traforest})}

\item{applyfun}{(see \code{?trtf::predict.traforest})}

\item{cores}{(see \code{?trtf::predict.traforest})}

\item{...}{(see \code{?trtf::predict.traforest})}
}
\description{
Transformation Forests
See \code{?trtf::predict.traforest}.
}
