% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchObjectForValue.R
\name{searchObjectForValue}
\alias{searchObjectForValue}
\title{Search for value in complex R objects}
\usage{
searchObjectForValue(obj, value, show_output = FALSE)
}
\arguments{
\item{obj}{[`R object`] R object for which we search for `value`.}

\item{value}{[`numeric(1L)|character(1L)`] Value which which for which `obj` is searched through.}

\item{show_output}{[`logical(1L)`] Show log (default `show_output = FALSE`).}
}
\description{
Search for value in complex R objects
}
\examples{
obj = lm(Petal.Length ~ ., data = iris)
lfound = searchObjectForValue(obj, iris$Sepal.Length[1])
# Get values found:
lapply(lfound$found, function(l) eval(parse(text = l)))
}
