% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traforest.R
\name{traforest}
\alias{traforest}
\title{Transformation Forests
See \code{?trtf::traforest}.}
\usage{
traforest(
  object,
  parm = 1:length(coef(object)),
  reparm = NULL,
  update = TRUE,
  min_update = length(coef(object)) * 2,
  mltargs = list(maxit = 10000),
  ...
)
}
\arguments{
\item{object}{(see \code{?trtf::traforest})}

\item{parm}{(see \code{?trtf::traforest})}

\item{reparm}{(see \code{?trtf::traforest})}

\item{update}{(see \code{?trtf::traforest})}

\item{min_update}{(see \code{?trtf::traforest})}

\item{mltargs}{(see \code{?trtf::traforest})}

\item{...}{(see \code{?trtf::traforest})}
}
\description{
Transformation Forests
See \code{?trtf::traforest}.
}
