% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructDataSkeleton.R
\name{constructDataSkeleton}
\alias{constructDataSkeleton}
\title{Construct skeleton of a data set}
\usage{
constructDataSkeleton(dat, n = NULL, sm = 0.01)
}
\arguments{
\item{dat}{\code{\link[=data.frame]{data.frame()}} Data set for which the skeleton is calculated}

\item{n}{\code{\link{integer(1L)}} Number of observations of the skeleton. The default equals the maximal
number of factor levels.}

\item{sm}{\code{\link{numeric(1L)}} Fraction of the range used as safety margin (default \code{sm = 0.01}).}
}
\value{
Skeleton \code{data.frame}
}
\description{
Representing the structure of a data set without including any observation of the data frame.
Numerical values are uniformly drawn between \code{min(x) - sm} and \code{max(x) + sm} where \code{sm} is
a "safety margin" to not get insights about the minimum and maximum of a feature. Categorical
features first contains the factor levels to ensure all levels in the skeleton. The remaining
observations are then filled by randomly sampling from the factor levels.
}
