% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchObjectForValue.R
\name{findNumber}
\alias{findNumber}
\title{Find number in R object}
\usage{
findNumber(obj, number)
}
\arguments{
\item{obj}{[`R object`] R object which should be flatten.}

\item{number}{[`anything`] Value which which for which `obj` is searched through.}
}
\value{
[`list`] Containing messages whether the object was found or not.
}
\description{
Find number in R object
}
